#yh-issue-2, + key event '<, >, shift'
