+++ .gitignore
@@ -4,3 +4,4 @@ package-lock.json
 electron-builder.yml
 
 resources/mock
+resources/image
