+++ src/index.js
@@ -1,5 +1,4 @@
-const drop = require("./renderer/drop");
-const filePath = require("./renderer/filePath");
+const drop = require("./util/drop");
 
 const csvMaker = require("./util/csvMaker");
 const keyboardControl = require("./util/keyboardControl");
